+++ CHANGELOG.md
@@ -4,9 +4,13 @@ This change log follows the conventions of [keepachangelog.com](http://keepachan
 
 ## [Unreleased]
 
+### Added 
+- [Simple sound alert](https://freesound.org/people/themusicalnomad/sounds/253886/) when pomodoro is stopped.
+
 ### Fixed
 - Fixed frequent application crashes caused by jnativehook library used for registering global keyboard shortcuts.
 [JKeyMaster](https://github.com/tulskiy/jkeymaster) is now used instead.
+
 
 
 [Unreleased]: https://github.com/your-name/change/compare/0.1.0...HEAD